requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
openai>=2.0.0
httpx==0.27.2
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
google-cloud-storage==2.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
numpy==1.24.3
scikit-learn==1.3.2
email-validator==2.1.0




